rename js-api-wrappers to substrate-api
