+++ lib/mocks/api/apiAssets.dart
@@ -1,6 +1,6 @@
 import 'package:encointer_wallet/mocks/data/mockAccountData.dart';
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
-import 'package:encointer_wallet/service/js_api_wrappers/apiAssets.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
+import 'package:encointer_wallet/service/substrate_api/apiAssets.dart';
 import 'package:encointer_wallet/store/app.dart';
 
 class MockApiAssets extends ApiAssets {
