+++ lib/page/account/import/importAccountPage.dart
@@ -1,6 +1,6 @@
 import 'package:encointer_wallet/page/account/create/createPinForm.dart';
 import 'package:encointer_wallet/page/account/import/importAccountForm.dart';
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/utils/format.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
