+++ lib/page/profile/settings/remoteNodeListPage.dart
@@ -1,5 +1,5 @@
 import 'package:encointer_wallet/config/consts.dart';
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/store/settings.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
 import 'package:encointer_wallet/utils/translations/translations.dart';
