+++ lib/service/substrate_api/apiAssets.dart
@@ -1,4 +1,4 @@
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/store/app.dart';
 
 class ApiAssets {
