+++ lib/service/substrate_api/codecApi.dart
@@ -1,7 +1,7 @@
 import 'dart:convert';
 import 'dart:typed_data';
 
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 
 const String ClaimOfAttendanceJSRegistryName = 'ClaimOfAttendance';
