+++ lib/service/substrate_api/encointer/apiTeeProxy.dart
@@ -1,6 +1,6 @@
 import 'dart:convert';
 
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/store/encointer/types/communities.dart';
 import 'package:encointer_wallet/store/encointer/types/encointerBalanceData.dart';
 import 'package:encointer_wallet/store/encointer/types/workerApi.dart';
