+++ lib/service/substrate_api/types/genExternalLinksParams.dart
