+++ lib/utils/UI.dart
@@ -3,7 +3,7 @@ import 'dart:io';
 
 import 'package:encointer_wallet/common/regInputFormatter.dart';
 import 'package:encointer_wallet/config/consts.dart';
-import 'package:encointer_wallet/service/js_api_wrappers/api.dart';
+import 'package:encointer_wallet/service/substrate_api/api.dart';
 import 'package:encointer_wallet/service/udpateJSCodeApi.dart';
 import 'package:encointer_wallet/store/app.dart';
 import 'package:encointer_wallet/utils/translations/index.dart';
